Update to convention v1
