+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -8,13 +8,13 @@ assignees: ''
 ---
 
 **Describe the bug**
-A clear and concise description of what the bug is.
+A clear and concise description of the bug.
 
-**To Reproduce**
+**Steps to reproduce**
 Steps to reproduce the behavior:
 1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
+2. Run '...'
+3. Provide '...'
 4. See error
 
 **Expected behavior**
@@ -23,9 +23,9 @@ A clear and concise description of what you expected to happen.
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
 
-**Desktop (please complete the following information):**
+**Environment (please complete the following information):**
  - OS: [e.g. Linux Fedora 43, Windows 11...]
- - Version [e.g. v0.1.1]
+ - Version: [e.g. v0.1.1]
 
 **Additional context**
 Add any other context about the problem here.
