--- .github/ISSUE_TEMPLATE/config.yaml
@@ -1,5 +1,0 @@
-blank_issues_enabled: false
-contact_links:
-  - name: General information
-    url: https://www.modelblocks.org/
-    about: Please consult our website for general information on the Modelblocks methodology.
