+++ .pre-commit-config.yaml
@@ -25,15 +25,15 @@ repos:
 
   # Python file formatting
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.15.9
+    rev: v0.15.18
     hooks:
-      - id: ruff
+      - id: ruff-check
         args: [--fix, --exit-non-zero-on-fix]
       - id: ruff-format
 
   # Snakemake file formatting
   - repo: https://github.com/snakemake/snakefmt
-    rev: v1.0.0
+    rev: v2.0.2
     hooks:
       - id: snakefmt
