+++ AUTHORS
@@ -2,6 +2,6 @@ This is the list of contributors for copyright purposes.
 
 This does not necessarily list everyone who has contributed to this software's
 code or documentation. For a full contributor list, see:
-<https://github.com/calliope-project/module_powerplants/graphs/contributors>
+<https://github.com/modelblocks-org/module_powerplants/graphs/contributors>
 
 Ivan Ruiz Manuel, <i.ruizmanuel@tudelft.nl>
