+++ INTERFACE.yaml
@@ -1,15 +1,17 @@
-# Module Input-Output structure for automated doc. generation
+# Module Input-Output structure for automated docs generation
+convention_version: v1.0.0
+
 pathvars:
   snakemake_defaults:
     logs:
       default: "<logs>"
-      description: location of rule log files.
+      description: location of snakemake log files.
     resources:
       default: "<resources>"
-      description: "location of module resource files."
+      description: "location of module resource (input) files."
     results:
       default: "<results>"
-      description: "location of module results."
+      description: "location of module result (output) files."
   user_resources:
     shapes:
       default: "<resources>/user/{shapes}/shapes.parquet"
