+++ README.md
@@ -1,6 +1,6 @@
-# Modelblocks - Powerplants module
+# Powerplants module
 
-Aggregate global powerplant capacities into any resolution.
+A data module to estimate global powerplant capacities for any region in the world at any resolution.
 
 <!-- Place an attractive image of module outputs here -->
 <p align="center">
@@ -101,6 +101,21 @@ cd module_powerplants
 pixi install --all
 ```
 
+Please be aware that this is a multi-environment project (see [pixi.toml](./pixi.toml) for details).
+- `default`: used for development and integration testing.
+Because it contains `Snakemake`, `conda` and `pytest` as dependencies it **should not be used** in `Snakemake` rules.
+- `module`: contains minimal dependencies used in `Snakemake` rules.
+If modified, be sure to export it to `Snakemake` so it can be recreated by module users:
+
+```shell
+# create module.yaml and conda-spec pin files in workflow/envs/
+pixi run export-snakemake-env module
+```
+
+
+## Testing
+<!-- Please do not modify this templated section -->
+
 For testing, simply run:
 
 ```shell
@@ -118,7 +133,7 @@ snakemake --use-conda --cores 2  # run the workflow!
 ## References
 <!-- Please provide thorough referencing below -->
 
-This module is based on the following research and datasets.
+This module is based on the following research and datasets:
 For specific versions please consult our [stable dataset repository](https://doi.org/10.5281/zenodo.16037139).
 
 * **Global Energy Monitor datasets.** <https://globalenergymonitor.org/>. License: CC BY 4.0.
@@ -136,3 +151,17 @@ U.S. Energy Information Administration (Oct 2008). <https://www.eia.gov/internat
 * **Satellite Utility-scale PV dataset.**
 TransitionZero Solar Asset Mapper, TransitionZero. <https://www.transitionzero.org/products/solar-asset-mapper>.
 License: CC BY-NC 4.0.
+
+
+## Contributors ✨
+
+Thanks goes to these wonderful people, sorted alphabetically ([emoji key](https://allcontributors.org/en/reference/emoji-key/)):
+
+<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
+<!-- prettier-ignore-start -->
+<!-- markdownlint-disable -->
+<!-- markdownlint-restore -->
+<!-- prettier-ignore-end -->
+<!-- ALL-CONTRIBUTORS-LIST:END -->
+
+This project follows the [all-contributors](https://github.com/all-contributors/all-contributors) specification. Contributions of any kind welcome!
