+++ figures/.gitkeep
@@ -1,0 +1,1 @@
+# For module documentation figures
