+++ workflow/envs/module.yaml
@@ -1,0 +1,23 @@
+name: module
+channels:
+- conda-forge
+- bioconda
+- nodefaults
+dependencies:
+- curl >=8.9.1
+- click >=8.2.1
+- cmap >=0.7.2
+- dask >=2025.7.0
+- geopandas >=1.0.1
+- ipdb >=0.13.13
+- openpyxl >=3.1.5
+- numpy >=2.3.5
+- matplotlib >=3.10.3
+- pandas >=2.3.0
+- pandera-geopandas >=0.24.0
+- pyarrow >=19.0.1
+- pyproj >=3.7.1
+- python >=3.12.9
+- pyyaml >=6.0.2
+- xlrd >=2.0.1
+- gregor >=0.1.0
